+++ examples-pydantic-v1/equipment_unit_pydantic_v1/__init__.py
@@ -1,0 +1,3 @@
+from equipment_unit_pydantic_v1._api_client import EquipmentUnitClient
+
+__all__ = ["EquipmentUnitClient"]
