+++ examples-pydantic-v1/equipment_unit_pydantic_v1/_api/_core.py
@@ -16,7 +16,7 @@ from cognite.client.data_classes import TimeSeriesList
 from cognite.client.data_classes.data_modeling.instances import Instance
 from cognite.client.data_classes.data_modeling.instances import InstanceAggregationResultList
 
-from equipment_unit_pydantic_v1.client.data_classes._core import (
+from equipment_unit_pydantic_v1.data_classes._core import (
     DomainModel,
     DomainModelCore,
     DomainModelWrite,
@@ -33,7 +33,7 @@ from equipment_unit_pydantic_v1.client.data_classes._core import (
     DomainModelCore,
     DomainRelation,
 )
-from equipment_unit_pydantic_v1.client import data_classes
+from equipment_unit_pydantic_v1 import data_classes
 
 
 DEFAULT_LIMIT_READ = 25
