+++ examples-pydantic-v1/equipment_unit_pydantic_v1/_api/start_end_time.py
