+++ examples-pydantic-v1/equipment_unit_pydantic_v1/_api/unit_procedure.py
@@ -8,21 +8,24 @@ from cognite.client import CogniteClient
 from cognite.client import data_modeling as dm
 from cognite.client.data_classes.data_modeling.instances import InstanceAggregationResultList
 
-from equipment_unit_pydantic_v1.client.data_classes._core import DEFAULT_INSTANCE_SPACE
-from equipment_unit_pydantic_v1.client.data_classes import (
+from equipment_unit_pydantic_v1.data_classes._core import DEFAULT_INSTANCE_SPACE
+from equipment_unit_pydantic_v1.data_classes import (
     DomainModelCore,
     DomainModelWrite,
     ResourcesWriteResult,
-    WorkOrder,
-    WorkOrderWrite,
-    WorkOrderFields,
-    WorkOrderList,
-    WorkOrderWriteList,
-    WorkOrderTextFields,
+    UnitProcedure,
+    UnitProcedureWrite,
+    UnitProcedureFields,
+    UnitProcedureList,
+    UnitProcedureWriteList,
+    UnitProcedureTextFields,
+    StartEndTime,
+    StartEndTimeWrite,
+    StartEndTimeList,
 )
-from equipment_unit_pydantic_v1.client.data_classes._work_order import (
-    _WORKORDER_PROPERTIES_BY_FIELD,
-    _create_work_order_filter,
+from equipment_unit_pydantic_v1.data_classes._unit_procedure import (
+    _UNITPROCEDURE_PROPERTIES_BY_FIELD,
+    _create_unit_procedure_filter,
 )
 from ._core import (
     DEFAULT_LIMIT_READ,
@@ -33,79 +36,85 @@ from ._core import (
     QueryStep,
     QueryBuilder,
 )
-from .work_order_query import WorkOrderQueryAPI
+from .unit_procedure_work_orders import UnitProcedureWorkOrdersAPI
+from .unit_procedure_work_units import UnitProcedureWorkUnitsAPI
+from .unit_procedure_query import UnitProcedureQueryAPI
 
 
-class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
+class UnitProcedureAPI(NodeAPI[UnitProcedure, UnitProcedureWrite, UnitProcedureList]):
     def __init__(self, client: CogniteClient, view_by_read_class: dict[type[DomainModelCore], dm.ViewId]):
-        view_id = view_by_read_class[WorkOrder]
+        view_id = view_by_read_class[UnitProcedure]
         super().__init__(
             client=client,
             sources=view_id,
-            class_type=WorkOrder,
-            class_list=WorkOrderList,
-            class_write_list=WorkOrderWriteList,
+            class_type=UnitProcedure,
+            class_list=UnitProcedureList,
+            class_write_list=UnitProcedureWriteList,
             view_by_read_class=view_by_read_class,
         )
         self._view_id = view_id
+        self.work_orders_edge = UnitProcedureWorkOrdersAPI(
+            client, view_by_read_class, StartEndTime, StartEndTimeWrite, StartEndTimeList
+        )
+        self.work_units_edge = UnitProcedureWorkUnitsAPI(
+            client, view_by_read_class, StartEndTime, StartEndTimeWrite, StartEndTimeList
+        )
 
     def __call__(
         self,
-        description: str | list[str] | None = None,
-        description_prefix: str | None = None,
-        performed_by: str | list[str] | None = None,
-        performed_by_prefix: str | None = None,
+        name: str | list[str] | None = None,
+        name_prefix: str | None = None,
         type_: str | list[str] | None = None,
         type_prefix: str | None = None,
         external_id_prefix: str | None = None,
         space: str | list[str] | None = None,
         limit: int | None = DEFAULT_QUERY_LIMIT,
         filter: dm.Filter | None = None,
-    ) -> WorkOrderQueryAPI[WorkOrderList]:
-        """Query starting at work orders.
+    ) -> UnitProcedureQueryAPI[UnitProcedureList]:
+        """Query starting at unit procedures.
 
         Args:
-            description: The description to filter on.
-            description_prefix: The prefix of the description to filter on.
-            performed_by: The performed by to filter on.
-            performed_by_prefix: The prefix of the performed by to filter on.
+            name: The name to filter on.
+            name_prefix: The prefix of the name to filter on.
             type_: The type to filter on.
             type_prefix: The prefix of the type to filter on.
             external_id_prefix: The prefix of the external ID to filter on.
             space: The space to filter on.
-            limit: Maximum number of work orders to return. Defaults to 25. Set to -1, float("inf") or None to return all items.
+            limit: Maximum number of unit procedures to return. Defaults to 25. Set to -1, float("inf") or None to return all items.
             filter: (Advanced) If the filtering available in the above is not sufficient, you can write your own filtering which will be ANDed with the filter above.
 
         Returns:
-            A query API for work orders.
+            A query API for unit procedures.
 
         """
         has_data = dm.filters.HasData(views=[self._view_id])
-        filter_ = _create_work_order_filter(
+        filter_ = _create_unit_procedure_filter(
             self._view_id,
-            description,
-            description_prefix,
-            performed_by,
-            performed_by_prefix,
+            name,
+            name_prefix,
             type_,
             type_prefix,
             external_id_prefix,
             space,
             (filter and dm.filters.And(filter, has_data)) or has_data,
         )
-        builder = QueryBuilder(WorkOrderList)
-        return WorkOrderQueryAPI(self._client, builder, self._view_by_read_class, filter_, limit)
+        builder = QueryBuilder(UnitProcedureList)
+        return UnitProcedureQueryAPI(self._client, builder, self._view_by_read_class, filter_, limit)
 
     def apply(
         self,
-        work_order: WorkOrderWrite | Sequence[WorkOrderWrite],
+        unit_procedure: UnitProcedureWrite | Sequence[UnitProcedureWrite],
         replace: bool = False,
         write_none: bool = False,
     ) -> ResourcesWriteResult:
-        """Add or update (upsert) work orders.
+        """Add or update (upsert) unit procedures.
+
+        Note: This method iterates through all nodes and timeseries linked to unit_procedure and creates them including the edges
+        between the nodes. For example, if any of `work_orders` or `work_units` are set, then these
+        nodes as well as any nodes linked to them, and all the edges linking these nodes will be created.
 
         Args:
-            work_order: Work order or sequence of work orders to upsert.
+            unit_procedure: Unit procedure or sequence of unit procedures to upsert.
             replace (bool): How do we behave when a property value exists? Do we replace all matching and existing values with the supplied values (true)?
                 Or should we merge in new values for properties together with the existing values (false)? Note: This setting applies for all nodes or edges specified in the ingestion call.
             write_none (bool): This method, will by default, skip properties that are set to None. However, if you want to set properties to None,
@@ -115,51 +124,51 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
 
         Examples:
 
-            Create a new work_order:
+            Create a new unit_procedure:
 
-                >>> from equipment_unit_pydantic_v1.client import EquipmentUnitClient
-                >>> from equipment_unit_pydantic_v1.client.data_classes import WorkOrderWrite
+                >>> from equipment_unit_pydantic_v1 import EquipmentUnitClient
+                >>> from equipment_unit_pydantic_v1.data_classes import UnitProcedureWrite
                 >>> client = EquipmentUnitClient()
-                >>> work_order = WorkOrderWrite(external_id="my_work_order", ...)
-                >>> result = client.work_order.apply(work_order)
+                >>> unit_procedure = UnitProcedureWrite(external_id="my_unit_procedure", ...)
+                >>> result = client.unit_procedure.apply(unit_procedure)
 
         """
         warnings.warn(
             "The .apply method is deprecated and will be removed in v1.0. "
             "Please use the .upsert method on the client instead. This means instead of "
-            "`my_client.work_order.apply(my_items)` please use `my_client.upsert(my_items)`."
+            "`my_client.unit_procedure.apply(my_items)` please use `my_client.upsert(my_items)`."
             "The motivation is that all apply methods are the same, and having one apply method per API "
             " class encourages users to create items in small batches, which is inefficient."
             "In addition, .upsert method is more descriptive of what the method does.",
             UserWarning,
             stacklevel=2,
         )
-        return self._apply(work_order, replace, write_none)
+        return self._apply(unit_procedure, replace, write_none)
 
     def delete(
         self, external_id: str | SequenceNotStr[str], space: str = DEFAULT_INSTANCE_SPACE
     ) -> dm.InstancesDeleteResult:
-        """Delete one or more work order.
+        """Delete one or more unit procedure.
 
         Args:
-            external_id: External id of the work order to delete.
-            space: The space where all the work order are located.
+            external_id: External id of the unit procedure to delete.
+            space: The space where all the unit procedure are located.
 
         Returns:
             The instance(s), i.e., nodes and edges which has been deleted. Empty list if nothing was deleted.
 
         Examples:
 
-            Delete work_order by id:
+            Delete unit_procedure by id:
 
-                >>> from equipment_unit_pydantic_v1.client import EquipmentUnitClient
+                >>> from equipment_unit_pydantic_v1 import EquipmentUnitClient
                 >>> client = EquipmentUnitClient()
-                >>> client.work_order.delete("my_work_order")
+                >>> client.unit_procedure.delete("my_unit_procedure")
         """
         warnings.warn(
             "The .delete method is deprecated and will be removed in v1.0. "
             "Please use the .delete method on the client instead. This means instead of "
-            "`my_client.work_order.delete(my_ids)` please use `my_client.delete(my_ids)`."
+            "`my_client.unit_procedure.delete(my_ids)` please use `my_client.delete(my_ids)`."
             "The motivation is that all delete methods are the same, and having one delete method per API "
             " class encourages users to delete items in small batches, which is inefficient.",
             UserWarning,
@@ -168,90 +177,104 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
         return self._delete(external_id, space)
 
     @overload
-    def retrieve(self, external_id: str, space: str = DEFAULT_INSTANCE_SPACE) -> WorkOrder | None: ...
+    def retrieve(self, external_id: str, space: str = DEFAULT_INSTANCE_SPACE) -> UnitProcedure | None: ...
 
     @overload
-    def retrieve(self, external_id: SequenceNotStr[str], space: str = DEFAULT_INSTANCE_SPACE) -> WorkOrderList: ...
+    def retrieve(self, external_id: SequenceNotStr[str], space: str = DEFAULT_INSTANCE_SPACE) -> UnitProcedureList: ...
 
     def retrieve(
         self, external_id: str | SequenceNotStr[str], space: str = DEFAULT_INSTANCE_SPACE
-    ) -> WorkOrder | WorkOrderList | None:
-        """Retrieve one or more work orders by id(s).
+    ) -> UnitProcedure | UnitProcedureList | None:
+        """Retrieve one or more unit procedures by id(s).
 
         Args:
-            external_id: External id or list of external ids of the work orders.
-            space: The space where all the work orders are located.
+            external_id: External id or list of external ids of the unit procedures.
+            space: The space where all the unit procedures are located.
 
         Returns:
-            The requested work orders.
+            The requested unit procedures.
 
         Examples:
 
-            Retrieve work_order by id:
+            Retrieve unit_procedure by id:
 
-                >>> from equipment_unit_pydantic_v1.client import EquipmentUnitClient
+                >>> from equipment_unit_pydantic_v1 import EquipmentUnitClient
                 >>> client = EquipmentUnitClient()
-                >>> work_order = client.work_order.retrieve("my_work_order")
+                >>> unit_procedure = client.unit_procedure.retrieve("my_unit_procedure")
 
         """
-        return self._retrieve(external_id, space)
+        return self._retrieve(
+            external_id,
+            space,
+            retrieve_edges=True,
+            edge_api_name_type_direction_view_id_penta=[
+                (
+                    self.work_orders_edge,
+                    "work_orders",
+                    dm.DirectRelationReference("IntegrationTestsImmutable", "UnitProcedure.work_order"),
+                    "outwards",
+                    dm.ViewId("IntegrationTestsImmutable", "WorkOrder", "c5543fb2b1bc81"),
+                ),
+                (
+                    self.work_units_edge,
+                    "work_units",
+                    dm.DirectRelationReference("IntegrationTestsImmutable", "UnitProcedure.equipment_module"),
+                    "outwards",
+                    dm.ViewId("IntegrationTestsImmutable", "EquipmentModule", "b1cd4bf14a7a33"),
+                ),
+            ],
+        )
 
     def search(
         self,
         query: str,
-        properties: WorkOrderTextFields | Sequence[WorkOrderTextFields] | None = None,
-        description: str | list[str] | None = None,
-        description_prefix: str | None = None,
-        performed_by: str | list[str] | None = None,
-        performed_by_prefix: str | None = None,
+        properties: UnitProcedureTextFields | Sequence[UnitProcedureTextFields] | None = None,
+        name: str | list[str] | None = None,
+        name_prefix: str | None = None,
         type_: str | list[str] | None = None,
         type_prefix: str | None = None,
         external_id_prefix: str | None = None,
         space: str | list[str] | None = None,
         limit: int | None = DEFAULT_LIMIT_READ,
         filter: dm.Filter | None = None,
-    ) -> WorkOrderList:
-        """Search work orders
+    ) -> UnitProcedureList:
+        """Search unit procedures
 
         Args:
             query: The search query,
             properties: The property to search, if nothing is passed all text fields will be searched.
-            description: The description to filter on.
-            description_prefix: The prefix of the description to filter on.
-            performed_by: The performed by to filter on.
-            performed_by_prefix: The prefix of the performed by to filter on.
+            name: The name to filter on.
+            name_prefix: The prefix of the name to filter on.
             type_: The type to filter on.
             type_prefix: The prefix of the type to filter on.
             external_id_prefix: The prefix of the external ID to filter on.
             space: The space to filter on.
-            limit: Maximum number of work orders to return. Defaults to 25. Set to -1, float("inf") or None to return all items.
+            limit: Maximum number of unit procedures to return. Defaults to 25. Set to -1, float("inf") or None to return all items.
             filter: (Advanced) If the filtering available in the above is not sufficient, you can write your own filtering which will be ANDed with the filter above.
 
         Returns:
-            Search results work orders matching the query.
+            Search results unit procedures matching the query.
 
         Examples:
 
-           Search for 'my_work_order' in all text properties:
+           Search for 'my_unit_procedure' in all text properties:
 
-                >>> from equipment_unit_pydantic_v1.client import EquipmentUnitClient
+                >>> from equipment_unit_pydantic_v1 import EquipmentUnitClient
                 >>> client = EquipmentUnitClient()
-                >>> work_orders = client.work_order.search('my_work_order')
+                >>> unit_procedures = client.unit_procedure.search('my_unit_procedure')
 
         """
-        filter_ = _create_work_order_filter(
+        filter_ = _create_unit_procedure_filter(
             self._view_id,
-            description,
-            description_prefix,
-            performed_by,
-            performed_by_prefix,
+            name,
+            name_prefix,
             type_,
             type_prefix,
             external_id_prefix,
             space,
             filter,
         )
-        return self._search(self._view_id, query, _WORKORDER_PROPERTIES_BY_FIELD, properties, filter_, limit)
+        return self._search(self._view_id, query, _UNITPROCEDURE_PROPERTIES_BY_FIELD, properties, filter_, limit)
 
     @overload
     def aggregate(
@@ -262,14 +285,12 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
             | Sequence[Aggregations]
             | Sequence[dm.aggregations.MetricAggregation]
         ),
-        property: WorkOrderFields | Sequence[WorkOrderFields] | None = None,
+        property: UnitProcedureFields | Sequence[UnitProcedureFields] | None = None,
         group_by: None = None,
         query: str | None = None,
-        search_properties: WorkOrderTextFields | Sequence[WorkOrderTextFields] | None = None,
-        description: str | list[str] | None = None,
-        description_prefix: str | None = None,
-        performed_by: str | list[str] | None = None,
-        performed_by_prefix: str | None = None,
+        search_properties: UnitProcedureTextFields | Sequence[UnitProcedureTextFields] | None = None,
+        name: str | list[str] | None = None,
+        name_prefix: str | None = None,
         type_: str | list[str] | None = None,
         type_prefix: str | None = None,
         external_id_prefix: str | None = None,
@@ -287,14 +308,12 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
             | Sequence[Aggregations]
             | Sequence[dm.aggregations.MetricAggregation]
         ),
-        property: WorkOrderFields | Sequence[WorkOrderFields] | None = None,
-        group_by: WorkOrderFields | Sequence[WorkOrderFields] = None,
+        property: UnitProcedureFields | Sequence[UnitProcedureFields] | None = None,
+        group_by: UnitProcedureFields | Sequence[UnitProcedureFields] = None,
         query: str | None = None,
-        search_properties: WorkOrderTextFields | Sequence[WorkOrderTextFields] | None = None,
-        description: str | list[str] | None = None,
-        description_prefix: str | None = None,
-        performed_by: str | list[str] | None = None,
-        performed_by_prefix: str | None = None,
+        search_properties: UnitProcedureTextFields | Sequence[UnitProcedureTextFields] | None = None,
+        name: str | list[str] | None = None,
+        name_prefix: str | None = None,
         type_: str | list[str] | None = None,
         type_prefix: str | None = None,
         external_id_prefix: str | None = None,
@@ -311,14 +330,12 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
             | Sequence[Aggregations]
             | Sequence[dm.aggregations.MetricAggregation]
         ),
-        property: WorkOrderFields | Sequence[WorkOrderFields] | None = None,
-        group_by: WorkOrderFields | Sequence[WorkOrderFields] | None = None,
+        property: UnitProcedureFields | Sequence[UnitProcedureFields] | None = None,
+        group_by: UnitProcedureFields | Sequence[UnitProcedureFields] | None = None,
         query: str | None = None,
-        search_property: WorkOrderTextFields | Sequence[WorkOrderTextFields] | None = None,
-        description: str | list[str] | None = None,
-        description_prefix: str | None = None,
-        performed_by: str | list[str] | None = None,
-        performed_by_prefix: str | None = None,
+        search_property: UnitProcedureTextFields | Sequence[UnitProcedureTextFields] | None = None,
+        name: str | list[str] | None = None,
+        name_prefix: str | None = None,
         type_: str | list[str] | None = None,
         type_prefix: str | None = None,
         external_id_prefix: str | None = None,
@@ -326,7 +343,7 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
         limit: int | None = DEFAULT_LIMIT_READ,
         filter: dm.Filter | None = None,
     ) -> list[dm.aggregations.AggregatedNumberedValue] | InstanceAggregationResultList:
-        """Aggregate data across work orders
+        """Aggregate data across unit procedures
 
         Args:
             aggregate: The aggregation to perform.
@@ -334,15 +351,13 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
             group_by: The property to group by when doing the aggregation.
             query: The query to search for in the text field.
             search_property: The text field to search in.
-            description: The description to filter on.
-            description_prefix: The prefix of the description to filter on.
-            performed_by: The performed by to filter on.
-            performed_by_prefix: The prefix of the performed by to filter on.
+            name: The name to filter on.
+            name_prefix: The prefix of the name to filter on.
             type_: The type to filter on.
             type_prefix: The prefix of the type to filter on.
             external_id_prefix: The prefix of the external ID to filter on.
             space: The space to filter on.
-            limit: Maximum number of work orders to return. Defaults to 25. Set to -1, float("inf") or None to return all items.
+            limit: Maximum number of unit procedures to return. Defaults to 25. Set to -1, float("inf") or None to return all items.
             filter: (Advanced) If the filtering available in the above is not sufficient, you can write your own filtering which will be ANDed with the filter above.
 
         Returns:
@@ -350,20 +365,18 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
 
         Examples:
 
-            Count work orders in space `my_space`:
+            Count unit procedures in space `my_space`:
 
-                >>> from equipment_unit_pydantic_v1.client import EquipmentUnitClient
+                >>> from equipment_unit_pydantic_v1 import EquipmentUnitClient
                 >>> client = EquipmentUnitClient()
-                >>> result = client.work_order.aggregate("count", space="my_space")
+                >>> result = client.unit_procedure.aggregate("count", space="my_space")
 
         """
 
-        filter_ = _create_work_order_filter(
+        filter_ = _create_unit_procedure_filter(
             self._view_id,
-            description,
-            description_prefix,
-            performed_by,
-            performed_by_prefix,
+            name,
+            name_prefix,
             type_,
             type_prefix,
             external_id_prefix,
@@ -373,7 +386,7 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
         return self._aggregate(
             self._view_id,
             aggregate,
-            _WORKORDER_PROPERTIES_BY_FIELD,
+            _UNITPROCEDURE_PROPERTIES_BY_FIELD,
             property,
             group_by,
             query,
@@ -384,14 +397,12 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
 
     def histogram(
         self,
-        property: WorkOrderFields,
+        property: UnitProcedureFields,
         interval: float,
         query: str | None = None,
-        search_property: WorkOrderTextFields | Sequence[WorkOrderTextFields] | None = None,
-        description: str | list[str] | None = None,
-        description_prefix: str | None = None,
-        performed_by: str | list[str] | None = None,
-        performed_by_prefix: str | None = None,
+        search_property: UnitProcedureTextFields | Sequence[UnitProcedureTextFields] | None = None,
+        name: str | list[str] | None = None,
+        name_prefix: str | None = None,
         type_: str | list[str] | None = None,
         type_prefix: str | None = None,
         external_id_prefix: str | None = None,
@@ -399,34 +410,30 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
         limit: int | None = DEFAULT_LIMIT_READ,
         filter: dm.Filter | None = None,
     ) -> dm.aggregations.HistogramValue:
-        """Produces histograms for work orders
+        """Produces histograms for unit procedures
 
         Args:
             property: The property to use as the value in the histogram.
             interval: The interval to use for the histogram bins.
             query: The query to search for in the text field.
             search_property: The text field to search in.
-            description: The description to filter on.
-            description_prefix: The prefix of the description to filter on.
-            performed_by: The performed by to filter on.
-            performed_by_prefix: The prefix of the performed by to filter on.
+            name: The name to filter on.
+            name_prefix: The prefix of the name to filter on.
             type_: The type to filter on.
             type_prefix: The prefix of the type to filter on.
             external_id_prefix: The prefix of the external ID to filter on.
             space: The space to filter on.
-            limit: Maximum number of work orders to return. Defaults to 25. Set to -1, float("inf") or None to return all items.
+            limit: Maximum number of unit procedures to return. Defaults to 25. Set to -1, float("inf") or None to return all items.
             filter: (Advanced) If the filtering available in the above is not sufficient, you can write your own filtering which will be ANDed with the filter above.
 
         Returns:
             Bucketed histogram results.
 
         """
-        filter_ = _create_work_order_filter(
+        filter_ = _create_unit_procedure_filter(
             self._view_id,
-            description,
-            description_prefix,
-            performed_by,
-            performed_by_prefix,
+            name,
+            name_prefix,
             type_,
             type_prefix,
             external_id_prefix,
@@ -437,7 +444,7 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
             self._view_id,
             property,
             interval,
-            _WORKORDER_PROPERTIES_BY_FIELD,
+            _UNITPROCEDURE_PROPERTIES_BY_FIELD,
             query,
             search_property,
             limit,
@@ -446,53 +453,70 @@ class WorkOrderAPI(NodeAPI[WorkOrder, WorkOrderWrite, WorkOrderList]):
 
     def list(
         self,
-        description: str | list[str] | None = None,
-        description_prefix: str | None = None,
-        performed_by: str | list[str] | None = None,
-        performed_by_prefix: str | None = None,
+        name: str | list[str] | None = None,
+        name_prefix: str | None = None,
         type_: str | list[str] | None = None,
         type_prefix: str | None = None,
         external_id_prefix: str | None = None,
         space: str | list[str] | None = None,
         limit: int | None = DEFAULT_LIMIT_READ,
         filter: dm.Filter | None = None,
-    ) -> WorkOrderList:
-        """List/filter work orders
+        retrieve_edges: bool = True,
+    ) -> UnitProcedureList:
+        """List/filter unit procedures
 
         Args:
-            description: The description to filter on.
-            description_prefix: The prefix of the description to filter on.
-            performed_by: The performed by to filter on.
-            performed_by_prefix: The prefix of the performed by to filter on.
+            name: The name to filter on.
+            name_prefix: The prefix of the name to filter on.
             type_: The type to filter on.
             type_prefix: The prefix of the type to filter on.
             external_id_prefix: The prefix of the external ID to filter on.
             space: The space to filter on.
-            limit: Maximum number of work orders to return. Defaults to 25. Set to -1, float("inf") or None to return all items.
+            limit: Maximum number of unit procedures to return. Defaults to 25. Set to -1, float("inf") or None to return all items.
             filter: (Advanced) If the filtering available in the above is not sufficient, you can write your own filtering which will be ANDed with the filter above.
+            retrieve_edges: Whether to retrieve `work_orders` or `work_units` external ids for the unit procedures. Defaults to True.
 
         Returns:
-            List of requested work orders
+            List of requested unit procedures
 
         Examples:
 
-            List work orders and limit to 5:
+            List unit procedures and limit to 5:
 
-                >>> from equipment_unit_pydantic_v1.client import EquipmentUnitClient
+                >>> from equipment_unit_pydantic_v1 import EquipmentUnitClient
                 >>> client = EquipmentUnitClient()
-                >>> work_orders = client.work_order.list(limit=5)
+                >>> unit_procedures = client.unit_procedure.list(limit=5)
 
         """
-        filter_ = _create_work_order_filter(
+        filter_ = _create_unit_procedure_filter(
             self._view_id,
-            description,
-            description_prefix,
-            performed_by,
-            performed_by_prefix,
+            name,
+            name_prefix,
             type_,
             type_prefix,
             external_id_prefix,
             space,
             filter,
         )
-        return self._list(limit=limit, filter=filter_)
+
+        return self._list(
+            limit=limit,
+            filter=filter_,
+            retrieve_edges=retrieve_edges,
+            edge_api_name_type_direction_view_id_penta=[
+                (
+                    self.work_orders_edge,
+                    "work_orders",
+                    dm.DirectRelationReference("IntegrationTestsImmutable", "UnitProcedure.work_order"),
+                    "outwards",
+                    dm.ViewId("IntegrationTestsImmutable", "WorkOrder", "c5543fb2b1bc81"),
+                ),
+                (
+                    self.work_units_edge,
+                    "work_units",
+                    dm.DirectRelationReference("IntegrationTestsImmutable", "UnitProcedure.equipment_module"),
+                    "outwards",
+                    dm.ViewId("IntegrationTestsImmutable", "EquipmentModule", "b1cd4bf14a7a33"),
+                ),
+            ],
+        )
