--- examples-pydantic-v1/equipment_unit_pydantic_v1/client/__init__.py
@@ -1,3 +1,0 @@
-from equipment_unit_pydantic_v1.client._api_client import EquipmentUnitClient
-
-__all__ = ["EquipmentUnitClient"]
