+++ examples-pydantic-v1/equipment_unit_pydantic_v1/data_classes/_equipment_module.py
