+++ examples/windmill/data_classes/_core.py
@@ -466,7 +466,9 @@ T_DomainRelation = TypeVar("T_DomainRelation", bound=DomainRelation)
 
 
 def default_edge_external_id_factory(
-    start_node: DomainModelWrite | str, end_node: DomainModelWrite | str, edge_type: dm.DirectRelationReference
+    start_node: DomainModelWrite | str | dm.NodeId,
+    end_node: DomainModelWrite | str | dm.NodeId,
+    edge_type: dm.DirectRelationReference,
 ) -> str:
     start = start_node if isinstance(start_node, str) else start_node.external_id
     end = end_node if isinstance(end_node, str) else end_node.external_id
@@ -475,14 +477,17 @@ def default_edge_external_id_factory(
 
 class DomainRelationWrite(BaseModel, extra=Extra.forbid, populate_by_name=True):
     external_id_factory: ClassVar[
-        Callable[[Union[DomainModelWrite, str], Union[DomainModelWrite, str], dm.DirectRelationReference], str]
+        Callable[
+            [
+                Union[DomainModelWrite, str, dm.NodeId],
+                Union[DomainModelWrite, str, dm.NodeId],
+                dm.DirectRelationReference,
+            ],
+            str,
+        ]
     ] = default_edge_external_id_factory
     data_record: DataRecordWrite = Field(default_factory=DataRecordWrite)
     external_id: Optional[str] = Field(None, min_length=1, max_length=255)
-
-    @property
-    def data_records(self) -> DataRecordWriteList:
-        return DataRecordWriteList([node.data_record for node in self])
 
     @abstractmethod
     def _to_instances_write(
@@ -573,6 +578,10 @@ class DomainRelationWrite(BaseModel, extra=Extra.forbid, populate_by_name=True):
 
         return resources
 
+    @classmethod
+    def reset_external_id_factory(cls) -> None:
+        cls.external_id_factory = default_edge_external_id_factory
+
 
 T_DomainRelationWrite = TypeVar("T_DomainRelationWrite", bound=DomainRelationWrite)
 
@@ -582,6 +591,10 @@ class DomainRelationList(CoreList[T_DomainRelation]):
 
     def as_edge_ids(self) -> list[dm.EdgeId]:
         return [edge.as_id() for edge in self]
+
+    @property
+    def data_records(self) -> DataRecordWriteList:
+        return DataRecordWriteList([connection.data_record for connection in self])
 
 
 T_DomainRelationList = TypeVar("T_DomainRelationList", bound=DomainRelationList)
